requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pandas>=2.2.0
requests==2.31.0
schedule==1.2.0
pytz==2023.3
psutil>=5.9.0
